// src/store/enums/chats-action-type.enum.ts
export enum chatsActionType {
    chatListUpload = 'chats-chatList__upload',
    chatListUploaded = 'chats-chatList__uploaded',
    chatListUploadError = 'chats-chatList__uploadError',

    createChatUpload = 'chats-createChat__upload',
    createChatUploaded = 'chats-createChat__uploaded',
    createChatUploadError = 'chats-createChat__uploadError',

    deleteChatUpload = 'chats-deleteChat__upload',
    deleteChatUploaded = 'chats-deleteChat__uploaded',
    deleteChatUploadError = 'chats-deleteChat__uploadError',

    chatTokenUpload = 'chats-chatToken__upload',
    chatTokenUploaded = 'chats-chatToken__uploaded',
    chatTokenUploadError = 'chats-chatToken__uploadError',

    deleteChatUsersUpload = 'chats-deleteChatUsers__upload',
    deleteChatUsersUploaded = 'chats-deleteChatUsers__uploaded',
    deleteChatUsersUploadError = 'chats-deleteChatUsers__uploadError',

    addChatUsersUpload = 'chats-addChatUsers__upload',
    addChatUsersUploaded = 'chats-addChatUsers__uploaded',
    addChatUsersUploadError = 'chats-addChatUsers__uploadError',
}